requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
google-ads==22.1.0
google-api-core==2.14.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-cloud-secret-manager==2.16.4
google-crc32c==1.5.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.7
grpcio==1.59.3
grpcio-status==1.59.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
-e git+ssh://git@github.com/mateus-oliveira/google-ads-oauth2.git@cbe8275d07e2388a9e714d16d267c49cf1cc7988#egg=UNKNOWN&subdirectory=server
urllib3==2.1.0
Werkzeug==3.0.1
